% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wildfinder.R
\name{data.eco.list}
\alias{data.eco.list}
\title{Import WWF ecoregions data}
\usage{
data.eco.list()
}
\value{
a tibble containing ecoregions id and caracteristics
}
\description{
Import WWF ecoregions data
}
