% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wildfinder.R
\name{data.sp.eco.list}
\alias{data.sp.eco.list}
\title{Import WWF data linking ecoregions and mammal species}
\usage{
data.sp.eco.list()
}
\value{
a tibble containing ecoregions ids and mammal species ids
}
\description{
Import WWF data linking ecoregions and mammal species
}
