% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wildfinder.R
\name{data.sp.list}
\alias{data.sp.list}
\title{Import WWF species data}
\usage{
data.sp.list()
}
\value{
a tibble containing id of species and taxonomic information
}
\description{
Import WWF species data
}
