% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wildfinder.R
\name{data_sp_eco_list}
\alias{data_sp_eco_list}
\title{Import WWf species-ecoregions data}
\usage{
data_sp_eco_list()
}
\value{
A tibble liking species id to ecoregions id
}
\description{
Import WWf species-ecoregions data
}
