% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datawildfinder.R
\name{data_speeco_list}
\alias{data_speeco_list}
\title{Import WWF species-ecoregions Data}
\usage{
data_speeco_list()
}
\value{
A tibble containing ecoregions ID
}
\description{
Import WWF species-ecoregions Data
}
