% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wildfinder.R
\name{data_eco_list}
\alias{data_eco_list}
\title{Import WWF Ecoregions Data}
\usage{
data_eco_list()
}
\value{
A tibble containing ecoregion ID and geographical info
}
\description{
Import WWF Ecoregions Data
}
