% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wildfinder.R
\name{data_sp_eco}
\alias{data_sp_eco}
\title{Import WWF species-coregions data}
\usage{
data_sp_eco()
}
\value{
A tibble linking species IDs to ecoregions IDs
}
\description{
Import WWF species-coregions data
}
