% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wildfinder.R
\name{data_sp_eco_list}
\alias{data_sp_eco_list}
\title{Import WWF Ecoregions-Species Data}
\usage{
data_sp_eco_list()
}
\value{
A tibble linking ecoregion IDs and species ID
}
\description{
Import WWF Ecoregions-Species Data
}
