% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wildfinder.R
\name{data_sp_eco_list}
\alias{data_sp_eco_list}
\title{Import WWF species-ecoregions Data}
\usage{
data_sp_eco_list()
}
\value{
A tibble linking species ID to ecoregions ID
}
\description{
Import WWF species-ecoregions Data
}
